requirements.txt
aiogram==2.14.3
requests==2.26.0
pymongo==3.12.0
motor==2.4.0
peewee==3.14.4
psycopg2==2.9.1
python-dotenv==0.19.1